requirements.txt
appdirs==1.4.4
appnope==0.1.3
asttokens==2.4.0
attrs==23.1.0
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.2
boto3==1.28.62
botocore==1.31.62
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.3
cssselect==1.2.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
fake-useragent==1.2.1
filelock==3.12.4
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
ipython==8.15.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.0
jmespath==1.0.1
json5==0.9.14
lxml==4.9.3
matplotlib-inline==0.1.6
optional-django==0.1.0
outcome==1.2.0
packaging==23.1
parse==1.19.1
parsel==1.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
proces==0.1.7
prompt-toolkit==3.0.39
Protego==0.3.0
protobuf==4.21.12
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyee==8.2.2
Pygments==2.16.1
PyMySQL==1.1.0
pyOpenSSL==23.2.0
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
s3transfer==0.7.0
scrapeops-python-logger==0.4.7
scrapeops-python-requests==0.4.7
scrapeops-scrapy==0.5.3
Scrapy==2.10.1
scrapy-splash==0.9.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.2
tld==0.13
tldextract==3.5.0
tqdm==4.66.1
traitlets==5.10.0
trio==0.22.2
trio-websocket==0.10.4
Twisted==22.10.0
typing_extensions==4.7.1
urllib3==1.26.16
w3lib==2.1.2
wcwidth==0.2.6
websockets==10.4
wsproto==1.2.0
zipp==3.16.2
zope.interface==6.0
